+++ src/senaite/patient/browser/widgets/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .agedob import AgeDoBWidget  # noqa F401
+from .fullname import FullnameWidget  # noqa F401
+from .temporaryidentifier import TemporaryIdentifierWidget  # noqa F401